Version where struct are local
This version struct are not global but instead passed to the functions & threads that needs them. For optimization some of the variables should be uchar or defining a 1 bit flag should get better performance but I want to rework the code so that the only change would be replacing variables but a rework of the threads and how I achieve  race condition free consumer producer without mutex, spinlock  or semaphore.